requirements.txt
chainlit
httpx
azure-ai-inference